Moved all new classes/interfaces into "commands" package
