+++ lib/src/main/java/edu/touro/mco152/bm/App.java
@@ -4,7 +4,6 @@ package edu.touro.mco152.bm;
         import edu.touro.mco152.bm.ui.Gui;
         import edu.touro.mco152.bm.ui.MainFrame;
         import edu.touro.mco152.bm.ui.SelectFrame;
-
         import javax.swing.SwingWorker.StateValue;
         import javax.swing.*;
         import javax.swing.UIManager.LookAndFeelInfo;
